structure project correctly
