+++ pkg/twitter_auth/twitter_auth.go
@@ -1,0 +1,59 @@
+package twitterauth
+
+import (
+	"os"
+
+	"github.com/dghubble/go-twitter/twitter"
+	"github.com/dghubble/oauth1"
+	logr "github.com/sirupsen/logrus"
+)
+
+// Credentials struct contains API credentials pulled from env vars:
+type Credentials struct {
+	ConsumerKey       string
+	ConsumerSecret    string
+	AccessToken       string
+	AccessTokenSecret string
+}
+
+/* getCredentials:
+Input = N/A
+Return = Populated Credentials struct
+*/
+func GetCredentials() Credentials {
+	creds := Credentials{
+		AccessToken:       os.Getenv("ACCESS_TOKEN"),
+		AccessTokenSecret: os.Getenv("ACCESS_TOKEN_SECRET"),
+		ConsumerKey:       os.Getenv("CONSUMER_KEY"),
+		ConsumerSecret:    os.Getenv("CONSUMER_SECRET"),
+	}
+
+	return creds
+}
+
+/* GetUserClient:
+Input = credentials
+Return = authenticated twitter client, error
+*/
+func GetUserClient(creds *Credentials) (*twitter.Client, error) {
+
+	config := oauth1.NewConfig(creds.ConsumerKey, creds.ConsumerSecret)
+	token := oauth1.NewToken(creds.AccessToken, creds.AccessTokenSecret)
+
+	httpClient := config.Client(oauth1.NoContext, token)
+	client := twitter.NewClient(httpClient)
+
+	verifyParams := &twitter.AccountVerifyParams{
+		SkipStatus:   twitter.Bool(true),
+		IncludeEmail: twitter.Bool(true),
+	}
+
+	user, _, err := client.Accounts.VerifyCredentials(verifyParams)
+	if err != nil {
+		logr.Error(err)
+		return nil, err
+	}
+
+	logr.Infof("User Account Info:\n%+v\n", user)
+	return client, nil
+}
